corsan

+++ app/Console/Commands/UpdateInformation.php
@@ -68,14 +68,14 @@ class UpdateInformation extends Command {
             } elseif($config->ConecctionType == 'ws') {
                 $archivosPlanos = $this->importData($db);
             } elseif($config->ConecctionType == 'api'){
-                $token = $this->loginToApi($db);
+                $archivosPlanos = $this->loginToApi($config,$db);
             }
           
             // Function to configure and migrate tables (MigrateTrait).
             if($archivosPlanos == true){
                 $customMigrations = Custom_Migration::getAll();
                 foreach ($customMigrations as $migration) {
-                    if($migration->command == ":refresh"){
+                    if($migration->command == ":refresh" && $migration->custom_inserts_id != null){
                         DB::connection('dynamic_connection')->table($migration->name_table)->truncate();
                         print '◘ Tabla'.$migration->name_table." truncada.\n";
                     }
